invite_project/requirements.txt
asgiref==3.8.1
colorama==0.4.6
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
mysqlclient==2.2.4
pillow==10.3.0
PyJWT==2.8.0
pypng==0.20220715.0
qrcode==7.4.2
sqlparse==0.5.0
typing_extensions==4.12.0
tzdata==2024.1
